[NET-565][NET-568] Move base manager to single package
