+++ cmdb/framework/managers/error/__init__.py
@@ -13,15 +13,3 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-from cmdb.framework.managers.error.manager_errors import ManagerGetError, ManagerInsertError, ManagerUpdateError, \
-    ManagerDeleteError
-
-__all__ = [
-    ManagerGetError,
-    ManagerInsertError,
-    ManagerUpdateError,
-    ManagerDeleteError
-]
-
-
