+++ cmdb/manager/errors.py
